% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{generateReportTemplate}
\alias{generateReportTemplate}
\title{Function to generate report template using list object reportInfo}
\usage{
generateReportTemplate(reportInfo, report_config, coverage_data)
}
\arguments{
\item{reportInfo}{Named list from report builder/automatic generation tools with sample and report information}

\item{report_config}{Report config variables}

\item{coverage_data}{Report coverage table definitions to generate the coverage table}
}
\value{
report_template which is a word document template
}
\description{
Function to generate report template using list object reportInfo
}
