% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{getCoverageData}
\alias{getCoverageData}
\title{Generate sample gene coverage table (depending on the report type)}
\usage{
getCoverageData(seqrun, sample, path_gene_coverage_file, coverage_level)
}
\arguments{
\item{seqrun}{seqrun}

\item{sample}{sample name}

\item{path_gene_coverage_file}{file path to the run folders}

\item{coverage_level}{coverage level used}
}
\value{
report_template which is a word document template
}
\description{
Generate sample gene coverage table (depending on the report type)
}
