% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{getFusionInfo}
\alias{getFusionInfo}
\title{Fetch fusion info from pathos database
Used by Report Builder}
\usage{
getFusionInfo(con_pathOS, seqrun, sample_accession, report_config)
}
\arguments{
\item{con_pathOS}{PathOS DB connection}

\item{seqrun}{seqrun, character value}

\item{sample_accession}{sample accession, character value}

\item{report_config}{report config variables}
}
\value{
dataframe with fusion information
}
\description{
Fetch fusion info from pathos database
Used by Report Builder
}
