% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadAllReports}
\alias{loadAllReports}
\title{Function to load all reports from report builder database
Used by Report Browser}
\usage{
loadAllReports(con_rb)
}
\arguments{
\item{con_rb}{Report builder database connection}
}
\value{
All reports
}
\description{
Function to load all reports from report builder database
Used by Report Browser
}
