% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadClinicalContextMapping}
\alias{loadClinicalContextMapping}
\title{Load clinical context mapping data
Used by Report Builder, nvd identifier and report browser}
\usage{
loadClinicalContextMapping(con_rb)
}
\arguments{
\item{con_rb}{Report builder database connection}
}
\value{
clinical_context_mapping
}
\description{
Load clinical context mapping data
Used by Report Builder, nvd identifier and report browser
}
