% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadCoverageData}
\alias{loadCoverageData}
\title{Load coverage file data
Used by Report Builder, nvd identifier and report browser}
\usage{
loadCoverageData(con_rb)
}
\arguments{
\item{con_rb}{Report builder database connection}
}
\value{
coverage_data
}
\description{
Load coverage file data
Used by Report Builder, nvd identifier and report browser
}
