% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadPathOSData}
\alias{loadPathOSData}
\title{Load information in PathOS DB to prepare for saving a subset of samples of a seqrun
Used by NVD Identifier}
\usage{
loadPathOSData(con_pathOS, samples, seqrun, report_config)
}
\arguments{
\item{con_pathOS}{PathOS database connection}

\item{samples}{vector of sample accessions}

\item{seqrun}{seqrun}

\item{report_config}{report config variables}
}
\value{
a list of reportInfo objects one per sample (named list)
}
\description{
Load information in PathOS DB to prepare for saving a subset of samples of a seqrun
Used by NVD Identifier
}
