% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadReportBuilderData}
\alias{loadReportBuilderData}
\title{Load sample information to reportInfo object and report information to reportInfo objects
If there is a report existing, only load basic report information if there are multiple reports per sample
Used by Report Builder}
\usage{
loadReportBuilderData(con_rb, reportInfo)
}
\arguments{
\item{con_rb}{Report builder database connection}

\item{reportInfo}{Named list from report builder/automatic generation tools with sample and report information}
}
\value{
reportInfo
}
\description{
Load sample information to reportInfo object and report information to reportInfo objects
If there is a report existing, only load basic report information if there are multiple reports per sample
Used by Report Builder
}
