% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadReportBuilderInfo}
\alias{loadReportBuilderInfo}
\title{Check whether samples/reports exist in report builder database for the seqrun
Used by NVD Identifier}
\usage{
loadReportBuilderInfo(con_rb, seqrun)
}
\arguments{
\item{con_rb}{Report builder database connection}

\item{seqrun}{seqrun}
}
\value{
a data frame with Sample, Report and ReportBuilderInfo tables
}
\description{
Check whether samples/reports exist in report builder database for the seqrun
Used by NVD Identifier
}
