% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadReportBuilderReport}
\alias{loadReportBuilderReport}
\title{Load report information to reportInfo object
This call uses the reportID in reportInfo object
Used by Report Builder}
\usage{
loadReportBuilderReport(con_rb, reportInfo, report_config)
}
\arguments{
\item{con_rb}{Report builder database connection}

\item{reportInfo}{Named list from report builder/automatic generation tools with sample and report information}

\item{report_config}{report config variables}
}
\value{
reportInfo
}
\description{
Load report information to reportInfo object
This call uses the reportID in reportInfo object
Used by Report Builder
}
