% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadReportConfig}
\alias{loadReportConfig}
\title{Load report config information
Used by Report Builder, nvd identifier and report browser}
\usage{
loadReportConfig(con_rb)
}
\arguments{
\item{con_rb}{Report builder database connection}
}
\value{
report_config
}
\description{
Load report config information
Used by Report Builder, nvd identifier and report browser
}
