% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadSampleInfo}
\alias{loadSampleInfo}
\title{Load sample information to reportInfo object. This is the main reactive object used
by the report builder app which is also a named list
Used by Report Builder}
\usage{
loadSampleInfo(
  con_pathOS,
  seqrun,
  sample_accession,
  reportInfo,
  path_gene_coverage_file,
  report_config
)
}
\arguments{
\item{con_pathOS}{PathOS DB connection}

\item{seqrun}{seqrun, character value}

\item{sample_accession}{sample accession, character value}

\item{reportInfo}{Named list from report builder/automatic generation tools with report information}

\item{path_gene_coverage_file}{path to gene coverage file}

\item{report_config}{report config variables}
}
\value{
reportInfo
}
\description{
Load sample information to reportInfo object. This is the main reactive object used
by the report builder app which is also a named list
Used by Report Builder
}
