% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadSamples}
\alias{loadSamples}
\title{Load samples of the seqrun from PathOS
Used by NVD Identifier}
\usage{
loadSamples(con_pathOS, seqrun)
}
\arguments{
\item{con_pathOS}{PathOS database connection}

\item{seqrun}{seqrun}
}
\value{
a dataframe with sample information
}
\description{
Load samples of the seqrun from PathOS
Used by NVD Identifier
}
