% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadSNPBackbone}
\alias{loadSNPBackbone}
\title{Load SNP backbone data
Used by Report Builder, nvd identifier and report browser}
\usage{
loadSNPBackbone(con_rb)
}
\arguments{
\item{con_rb}{Report builder database connection}
}
\value{
snp_backbone
}
\description{
Load SNP backbone data
Used by Report Builder, nvd identifier and report browser
}
