% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadSpecimenLookup}
\alias{loadSpecimenLookup}
\title{Load Specimen lookup data for mapping Auslab data to specimen and specimen details
Used by Report Builder, nvd identifier and report browser}
\usage{
loadSpecimenLookup(con_rb)
}
\arguments{
\item{con_rb}{Report builder database connection}
}
\value{
specimen_lookup
}
\description{
Load Specimen lookup data for mapping Auslab data to specimen and specimen details
Used by Report Builder, nvd identifier and report browser
}
