% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadUserMapping}
\alias{loadUserMapping}
\title{Load user and display name mapping
Used by Report Builder, nvd identifier and report browser}
\usage{
loadUserMapping(con_rb)
}
\arguments{
\item{con_rb}{Report builder database connection}
}
\value{
user_mapping
}
\description{
Load user and display name mapping
Used by Report Builder, nvd identifier and report browser
}
