% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadVariants}
\alias{loadVariants}
\title{Load variants of the samples in seqrun from PathOS
Used by NVD Identifier}
\usage{
loadVariants(con_pathOS, samples, seqrun)
}
\arguments{
\item{con_pathOS}{PathOS database connection}

\item{samples}{vector of sample accessions}

\item{seqrun}{seqrun}
}
\value{
a dataframe with variant information
}
\description{
Load variants of the samples in seqrun from PathOS
Used by NVD Identifier
}
