% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{matchingPatientSamples}
\alias{matchingPatientSamples}
\title{Function to check whether there are other possible samples for the same patient}
\usage{
matchingPatientSamples(con_pathOS, reportInfo, report_config)
}
\arguments{
\item{con_pathOS}{PathOS DB connection}

\item{reportInfo}{Named list from report builder/automatic generation tools with sample and report information}

\item{report_config}{report config variables}
}
\value{
data frame of matching patient name, sample name and collection dates
}
\description{
Function to check whether there are other possible samples for the same patient
}
