% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\docType{data}
\name{report_config}
\alias{report_config}
\title{Report config file}
\format{
An object of class \code{list} of length 56.
}
\usage{
report_config
}
\description{
Report config file
}
\keyword{datasets}
