% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{saveNVDReports}
\alias{saveNVDReports}
\title{Save NVD reports to report builder db
Used by NVD Identifier}
\usage{
saveNVDReports(con_rb, report_DB_data, seqrun)
}
\arguments{
\item{con_rb}{Report builder database connection}

\item{report_DB_data}{data with NVD report information, one row per report, all tables in one row}

\item{seqrun}{seqrun}
}
\description{
Save NVD reports to report builder db
Used by NVD Identifier
}
