% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{saveSampleInfo}
\alias{saveSampleInfo}
\title{Save sample information from data file to report builder db
Used by NVD Identifier}
\usage{
saveSampleInfo(con_rb, data)
}
\arguments{
\item{con_rb}{Report builder database connection}

\item{data}{data with sample information}
}
\description{
Save sample information from data file to report builder db
Used by NVD Identifier
}
