% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{generateReportTemplate}
\alias{generateReportTemplate}
\title{Function to generate report template using list object reportInfo}
\usage{
generateReportTemplate(reportInfo)
}
\arguments{
\item{reportInfo}{Named list from report builder/automatic generation tools with sample and report information}
}
\value{
report_template which is a word document template
}
\description{
Function to generate report template using list object reportInfo
}
