% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{getVariantInfo}
\alias{getVariantInfo}
\title{Fetch variant info from pathos database
Used by Report Builder}
\usage{
getVariantInfo(con_pathOS, seqrun, sample_accession)
}
\arguments{
\item{con_pathOS}{PathOS DB connection}

\item{seqrun}{seqrun, character value}

\item{sample_accession}{sample accession, character value}
}
\value{
dataframe with variant information
}
\description{
Fetch variant info from pathos database
Used by Report Builder
}
