% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{loadSampleInfo}
\alias{loadSampleInfo}
\title{Load sample information to reportInfo object. This is the main reactive object used
by the report builder app which is also a named list}
\usage{
loadSampleInfo(con_pathOS, seqrun, sample_accession, reportInfo)
}
\arguments{
\item{con_pathOS}{PathOS DB connection}

\item{seqrun}{seqrun, character value}

\item{sample_accession}{sample accession, character value}

\item{reportInfo}{Named list from report builder/automatic generation tools with report information}
}
\value{
reportInfo
}
\description{
Load sample information to reportInfo object. This is the main reactive object used
by the report builder app which is also a named list
}
