% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{report_config}
\alias{report_config}
\title{Report config file shared between report builder and NVD Identifier}
\format{
An object of class \code{list} of length 56.
}
\usage{
report_config
}
\description{
Report config file shared between report builder and NVD Identifier
}
\keyword{datasets}
