% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWriter.R
\name{saveReport}
\alias{saveReport}
\title{Save report to Report builder database, also fetching report id etc when saved if this is a new report}
\usage{
saveReport(con_rb, reportInfo)
}
\arguments{
\item{con_rb}{Report builder database connection}

\item{reportInfo}{Named list from report builder/automatic generation tools with sample and report information}
}
\value{
reportInfo
}
\description{
Save report to Report builder database, also fetching report id etc when saved if this is a new report
}
